Added unit tests for NbtIntArray/TAG_Int_Array. Fixed a couple serialization bugs, in NbtString and NbtIntArray.

+++ LibNbt/Tags/NbtIntArray.cs
@@ -52,7 +52,7 @@ namespace LibNbt {
 
 
         internal override void WriteTag( NbtWriter writeStream, bool writeName ) {
-            writeStream.Write( NbtTagType.ByteArray );
+            writeStream.Write( NbtTagType.IntArray );
             if( writeName ) {
                 if( Name == null ) throw new NullReferenceException( "Name is null" );
                 writeStream.Write( Name );
